+++ gobblin-core/src/main/java/gobblin/CountBasedLimiter.java
@@ -10,7 +10,7 @@
  * CONDITIONS OF ANY KIND, either express or implied.
  */
 
-package gobblin.runtime;
+package gobblin;
 
 import java.io.Closeable;
