+++ gobblin-runtime/src/main/java/gobblin/runtime/LimitingExtractorDecorator.java
@@ -16,6 +16,7 @@ import java.io.IOException;
 
 import com.google.common.io.Closer;
 
+import gobblin.Limiter;
 import gobblin.source.extractor.DataRecordException;
 import gobblin.source.extractor.Extractor;
 import gobblin.util.Decorator;
