+++ gobblin-runtime/src/test/java/gobblin/runtime/local/LocalJobLauncherTest.java
@@ -25,8 +25,8 @@ import org.testng.annotations.Test;
 import gobblin.configuration.ConfigurationKeys;
 import gobblin.metastore.FsStateStore;
 import gobblin.metastore.StateStore;
-import gobblin.runtime.BaseLimiterType;
-import gobblin.runtime.DefaultLimiterFactory;
+import gobblin.BaseLimiterType;
+import gobblin.DefaultLimiterFactory;
 import gobblin.runtime.JobLauncherTestHelper;
 import gobblin.runtime.JobState;
 import gobblin.writer.Destination;
